Insert Invoice

+++ src/com/app/bookshop/dto/CartDto.java
@@ -8,7 +8,7 @@ public class CartDto implements Serializable{
 	private ArrayList<BookDto> books;
 	
 	private Integer itemTotal;
-	private Double moneyTotal;
+	private Integer moneyTotal;
 	
 	public ArrayList<BookDto> getBooks() {
 		return books;
@@ -35,8 +35,8 @@ public class CartDto implements Serializable{
 		}
 		return total;
 	}
-	public Double getMoneyTotal(){
-		Double total=0.0;
+	public Integer getMoneyTotal(){
+		Integer total=0;
 		for(int i=0;i<books.size();i++){
 			total += books.get(i).getAmount()*books.get(i).getPrice();
 		}
